Fix DFU handler being unable to access flash
The QSPI flash can only be accessed from tile 0. Thus the DFUHandler
is moved to run on tile 0. The audio task is not involved in the DFU
process anymore, apart from entering the dummy_deliver mode, where it
does not actually execute any audio i/o, but only serves the upwards
interface to avoid stalling the USB handler.

Signed-off-by: Julian Scheel <julian@jusst.de>