ECO2805: build 2.6.x with 40 kHz low pass filter
Stock firmware 2.6.0 has a 3rd stage knee at 8kHz to remove noise
from Vesper microphones. For critical infrastructure, we are using
an analog microphone that does not have the resonance at 12.5 kHz.
Build an alternative version of 2.6.x with a third-stage cutoff
at 40 kHz for high-frequency testing.

We will revert this code after building an xflashable version and a
DFUable version.