Allow filters to be selected at runtime instead of compile time.
Instead of loading directly from the constant pool, the code now loads a
register with the address of the desired filters. To minimize execution
time, filters are configured only at start. There are two switch-statement
equivalents that test each value of board rev and apply filters appropriately.

Code compiles but has not been tested.

Note the assembler does not take pure opcodes. It takes an instruction
plus an operating mode and coverts it into an opcode. It is
apparently expected that you can enter an opcode out of the
instruction set documentation and have the assembler reject it.

There are many quirky and weird limitation in the XS1 architecture.
Some that came up:

* stw can only load with an offset of 12, so the initial load of
  function argument in to r11[S_BOARDREV] failed because S_BOARDREV
  was added on the end of S_STORAGE and was too far away. Reassigned
  to apparently unused offset 3.
* I moved the filter load to before the `dualentsp 1` instruction
  to avoid the very gross need to add one to S_SETTINGS_OFFSET as
  a result of dualentsp messing with it. As I do not fully understand
  why dualentsp is needed---it adjusts the link register, but why
  do we need to do that?---this may not be correct
* lda cp[xxx] can only load into r11. Presumably this is because
  the assembler compiles:

ldaw r11, cp[g_second_stage_fir32]

into opcode ldwcpl, which is specific to r11.

   LDWCPL  -- Load word from large constant pool
   Loads a word relative to the constant pool pointer into r11. The offset can be larger
   than the offset specified in LDWCP. The instruction has one operand: op1 u20, a 20-bit
   immediate in the range 0...1048575. If u20 < 1024, the instruction requires no prefix.